requirements.txt
black==25.1.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
pillow_heif==0.22.0
platformdirs==4.3.7
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
Werkzeug==3.1.3
